Modify the rule of filtering glance tag
In the current progress of filtering glance image tags,
there are two problems:
1 When use "repo:tag" to pull image , if the glance image
tags were not updated , we will not get it even though the
glance image is what we want.
2 the current progress use update time to filter glance images,
this is an unreliable parameters which with some randomness.

we can use repo to find glance image firstly, if more than one
image is returned, then use tag to filter.

Change-Id: Ida4274688e7088e672c8963e75fd29ee524fbf97
Closes-bug: #1769912