Raise the right exception if network exists
This patch attempts to address the outstanding comments in
https://review.openstack.org/#/c/564800/17 .

If users create a network with an already exists neutron network,
Zun will raise an 'ContainerAlreadyExists' exception which is
confusing. This patch changes it to raise 'NetworkAlreadyExists'
instead. In addition, this patch also adds several unit test cases
for create_network.

Change-Id: I68e32143e0e750ab250a34ef0228a4778791e7a3