Change create_network from async to sync
The network create API will create the network in docker and
return immediately, so there is no need to perform a aysnc (202)
processing which has overhead on spawning a new thread for
each request.

This patch changes this API to a sync API. As a result, the response
code of create_network switches from 202 to 200. The RPC call changes
from 'cast' to 'call'. The compute manager will process the request
in main thread instead of spawning a new thread.

Change-Id: Ie1ec42e80f816e456372fb29db6ddaaa8a47c669