Prevent race condition on creating network
We leverage the unique constraint on network table to gurantee
at most one docker network can be created for one neutron network.
In particular, this patch moves the creation of network object from
API layer to the kuryr network driver, in which the docker network
is created. This basically ensures that each call to create docker
network is after a creation of network object and pass the unique
constraint imposed by the DB layer.

In addition, this patch also adds a destroy_network implementation
in objects and DB layer. We will clean up the DB record if the
network creation failed in docker.

Change-Id: I9507e90452c5af4cf1a64a045a07a9e63200b2b4
Closes-Bug: #1743498