Add exec_instances to data model
Zun is going to implement a proxy for interactive execute.
The first step is to introduce data model for tracking each
exec instances of a container. An exec instance in Zun is
an one-to-one mapping for exec instance in Docker.
We will use it to track a docker's exec instance as well as
a generated token for granting access of this exec instance.
In the future, the websocket proxy will retrieve the token from
incoming request and match it to the one stored in DB.
The request will be rejected if the token doesn't match.

This patch introduces the data model of exec_instance
in DB api and objects layer. It basically contains three
fields: container_id, exec_id, url, token.
The container_id is the id of the container record in Zun.
The exec_id is the ID of the docker's exec instance.
The url is the docker daemon endpoint for the exec instance
The token is as described above.

Partial-Bug: #1735076
Change-Id: Ib38a46c0e3f3aae58e1f562536b858bc4cd23bf8