Introduce 'fixed_ips' to network_list
Modify the response of network_list API. Add a 'fixed_ips' field
that contains the list of IP addresses and their subnet IDs.
This change is primarily for aligning the format with nova and
neutron so that API consumers can use the same parsing code.
The first use case is from Heat. Heat would be able to use the
same code for parsing nova's interface_list and zun's network_list.

Change-Id: Id1e652d944be852697f18958899e395918ab0885