Fix the json schema for network_attach
In before, the json schema didn't allow API calls for network_attach
with neither 'network' nor 'port' parameters. However, it is
a valid request to call network_attach without parameter. In this
case, Zun will choose a network for the container as how it is
done on container create.

This patch fixes the issue. With this patch, the network_attach API
(POST /containers/<container>/network_attach) will behave as following:
* Either 'network' or 'port' is provided -> ALLOW
* Neither 'network nor 'port' is provided -> ALLOW
* Both 'network' and 'port' is provided -> DISALLOW

Change-Id: I4933e14d850be10c16a1859cb8c4a8a09b75151d