Fall back to the legacy auth_uri option
In before, we switched from 'auth_uri' to 'www_authenticate_uri' [1]
in a backward-incompatible way, which breaks the legacy setup.
For example, if users install Zun with Pike or earlier version
of OpenStack, it will break.

This patch fixes it by fall back to the legacy option 'auth_uri'
if 'www_authenticate_uri' is not presented. In addition, a unnecessary
import_opt statement is removed.

[1] https://review.openstack.org/#/c/568457/

Change-Id: I5e3535f8dbfac73e56190442a31943ac22562ffb