Support file injection on container
Add support for file injection on creating the container.
API users need to pass the file contents by using the 'mounts'
attribute. The content will persist into a temp file in the
compute host and bind-mount into the container.

To achieve this, we introduce the volume driver called 'Local'.
This driver will implement the volume driver interface and
handle file injection related operations.

In data model, we adds a new field 'contents' into the volume_mapping
table. We also change 'volume_id' to be nullable because file
injection doesn't need a cinder volume.

A future work is to add the ability to limit the size and number
of injected files. The limits should be configurable either via
admin APIs or config options.

Implements: blueprint inject-files-to-container
Change-Id: I4ab6f50684f77bd7762e872d884ce11a7b0807ba