Verify docker API version if runtime is specified
The 'runtime' parameter is supported in Docker API version 1.26
or later. This patch add a validation on this requirement.

In addition, this patch allows Zun to work with legacy Docker
API version if 'runtime' parameter is not specified.
Right now, Kolla is using Docker API version 1.24 so it requires
this fix to make it work.

Change-Id: I04fb640977bfb6e0e5771806c9327b347fe8ed48
(cherry picked from commit c0ad682b005d9e0f27846329136dde3350cedf65)