Delete image residual data
When use“zun pull“ to pull image, the image is considered that
it can be managed by zun and the image is recorded in DB image
table. These image is allowed to be delete by client.

Those images that pulled when create container are considered
that they can not be managed by zun, because they were not
recorded in the DB image table. These images can be cleaned
by period task.

When delete the images managed by zun that pulled from glance,
delete the tar file and then delete the image loaded in docker.

Change-Id: I9d1e45af862dad26337047f5f6056d1d8aa5def8
Partial-Implements: blueprint delete-image-residual-data