Delete error volume which auto_remove is True
Create a container and create a volume at the same time. If the volume is
error , it would not be deleted even if it is auto_removed. Because this
error occured in _wait_for_volumes_available, VolumeMapping object has
not created in sql at that time. But the _fail_container get the volumes
from the sql.
  We can delete these volumes before calling _fail_container in
_wait_for_volumes_available().

Change-Id: I3cbf78bd57dcf56da09c94965679583fd63cdbd5
Closes-bug: #1783517