Disk size is smaller than base_device_size
When disk size is none and base_device_size is larger than default_disk,
the disk size is assigned to base_device_size.
Because base_device_size is float type, when converting to interger ,
the disk size is smaller than base_device_size.
We need to assign the ceil of  base_device_size to the disk size.

Change-Id: I5c3bfc7ce4c83c39a212040601792ff1ace2fdf5