Refactor list function in docker driver
Function list in docker driver is called by update_containers_states
in docker.driver and init_containers in compute.manager. When restart
compute service, the init_containers function is called, then the list
is called. The function list will heal non-existent containers, which
is not necessary to do this when restart compute service.
   It is better to heal non-existent containers only in
update_containers_states.

Change-Id: I677ca1000a48f0d5f768fcf9eacbd2a37522c9cb