Attached volume has no attached_servers
When use "cinder show" to look up the volume that
has been attached to a container, its status is "in-use"
but its "attached_servers" is none.

The reason is that Zun does not give a value of parameter
"instance_uuid" when call cinder interface of attach.
Then the container's uuid does not be recorded into volume's
attached_servers.

Change-Id: I3bd5e18f3f312a94cbbd4b3d2bbce3b0ae79ea69
Closes-bug: #1786162