Rework Cinder volume multi-attach
The current progress does not handle the volume residual information in
Cinder.

   For example, a volume attached to two containers which is on same or
different nodes. When detach the volume from the first container,we can
use "cinder show" to see the volume information. The attachment_id and
the attached_servers still have the first container's information.
  And then, detach volume from the second container, use "cinder show" to
see the volume information. The volume's status is still "in-use" and
attachment information is still existing.

    The reason is that when the volume is attached to more than one
containers, the current progress just destroyed the volume_mapping
table in Zun sql, and did nothing to Cinder.

    The cinder_api.detach should be called in any case, because it can
clean up attachment information in Cinder sql.

Co-Authored-By:  YaoZheng <zheng.yao1@zte.com.cn>
Change-Id: I05e29721a0dbe9d976fb8bc762ec8751dbaaf172
Closes-Bug: #1787308