Encode/Decode archive file data
In before, API users were not able to get/put archive file
that contains non-English charaters. This commit fixes
it by using the RFC 3548 encodings are suitable for encoding
binary data so that it can safely included as part of an HTTP
request/response. The archive file data will be encoded before
sending over HTTP and decoded after.

To maintain backward-compatibility, this commit also bumps the
API version to 1.25. API version in 1.24 or earlier won't
encode/decode the file data.

Needed-By: Id9fe25e06052d46dcc33a1e05100ef7a2d7db018
Change-Id: I1a2a9bada2f88d511b20fe586b9fa46b3a0289d6
Closes-Bug: #1789777