Fix a docker internal error on server restart
Zun creates docker network by passing the subnetpool and subnet
to kuryr-libnetwork. However, kuryr-libnetwork has an issue on
unseting neutron ports if a subnetpool is passed.
This commit avoid passing subnetpool when creating the docker
network.

Note: we passed subnetpool to kuryr in before because
kuryr doesn't take subnet as an input at that time so
we pass subnetpool for kuryr to search for the subnet.
Right now, kuryr support directly taking subnet as input so
subnetpool is not needed anymore.

Change-Id: I5a6ef58cd109749a4aac7e6295e0f6d737918ac0
Closes-Bug: #1791041