Use volume uuid based mountpoint
In before, we use the uuid of volume mapping as the basic to
compute the mountpoint. However, volume mapping represents
a mapping between a container and a local volume. If there is
multiple containers mapping to a volume, the computed mountpoint
is problematic because each container will compute a different
mountpoint.

This commits use the uuid of volume to compute the mountpoint.
As a result, in the shared volume case, all container will
compute the same mountpoint as long as they share the volume.

Change-Id: I0af30ecef043da000fea38776891feb19da018e6