Enable plugin in target branch
By default, devstack will enable plugin in master branch.
However, master branch might not be the branch we wanted.
In particular, if we are using stable branch, we want devstack
to enable plugin in stable branch instead of master branch.

This commit modifies the sample local conf to explicitly specify
the branch when enabling devstack plugin.

Change-Id: I1ce148f50e4d091e959b5c0d5b1b6281a07db946