Fix race condition on multinode
On a multinode environment, it is possible that multiple nodes
try to create a docker network at the same time. In before,
we didn't handle the race condition correctly so that
there are still concurrent requests on creating the docker
network. This commit refines the locking logic so that
it can handle the race condition correctly. In particular,
we let the first request to go through to create
the network and let other nodes to retry later.

Change-Id: Iaacc8285d2ea5170bcb57dfb449058359cd92c6b