Set binding:host_id on creating/updating port
This patch depends on a change on kuryr-libnetwork [1],
which skip populating 'binding:host_id' if it is already
populated. This commit leverages the kuryr's change and
has the 'binding:host_id' filled.

This effectively moves a neutron API call from kuryr
to zun but it increases the performance since docker
daemon are blocked if the slow API call is in kuryr side.

[1] https://review.openstack.org/#/c/627198/

Depends-On: https://review.openstack.org/#/c/627198/
Partial-Bug: #1809306
Change-Id: I67563ea069cafab16500bd8ba598743b7ae87fe6