Consolidate data model of Container and Capsule
In before, we defined different object for tracking Container
and Capsule respectively. The data model of Capsule and Container
objects are largely duplicated and hard to maintain.

This commit is mainly for consolidate the data model of Capsule
and Container. It defines a base class ContainerBase that
contains common code for all types of containers.
Then, it defines four subclass: Container, Capsule, CapsuleContainer,
and CapsuleInitContainer.

In database, all types of containers share a common 'container' table.
We added a column 'container_type' to indicate different types of
containers.

Change-Id: I9de6b213bf0c8541197e3d80b907fa3262c847e9