Fix capsule volume attach
In before, the schema 'requested_volume' parameter is different
between container and capsule. As the container and capsule
model being consolidated [1][2], it is better to unify the
schema of this parameter so that it is easier to maintain.

[1] https://review.openstack.org/#/c/636070/
[2] https://review.openstack.org/#/c/633371/

Closes-Bug: #1817488
Change-Id: Ifac6660230c926747ccfabb7d923f42825f30acc
(cherry picked from commit dd933447c024d0340c3d010da53bf61a723cd0cc)