Fix issue of sharing volume across containers
In a scenario that a cinder volume is attached to multiple containers
within a capsule, the compute manager did the cinder volume attach
multiple times for the same volume, which is not correct.
This commit fixes this issue.

Change-Id: Ic8d0c2056ee373a9c39803736eafb02b91c65bde
(cherry picked from commit ca85190e9da75cc5e85cf7bb01a8ad302b8209cd)