Fix issue on deleting cinder volume
Zun has a feature to auto cleanup container's volumes on deleting
a container. In before, compute manager issue API call to cinder
to request deleting the cinder volumes and wait for the cinder
to complete the deletion. However, we should wait for the volumes
that are auto-removed, because waiting for non-auto-removed volumes
will cause a timeout error obviously.

Change-Id: Idfb5429f618fc8163e87a7df7f8ed110f12ada1b