Fix lower-constraint deps handling
When the lower-constraints tox target was added, it was assumed the
install_command was just running the install and that the dependencies
and constraints were being set using "deps = ".

This fixed the install_command and deps to follow the expected pattern
so the lower-constraints job actual does install the lower constraints.

Raises the minimum for keystonemiddleware as the one listed was too low
for some of the changes merged to the code.

Change-Id: Ibbdc2d98272ceb65433818d8ba41ce3bfacd4443
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>