Support exposing ports in capsule - API
In capsule controller, process the 'ports' field and
convert it to 'exposed_ports' property of the container.
Right now, we process the 'containerPort' and 'protocol' field.
Under the hook, Zun will create a security group for opening
that port and protocol. The 'hostPort' field is ignored since
Zun doesn't expose compute host information. This might be revisited
later and implemented by the neutron floating IP port forwarding
feature.

Change-Id: Ic71d19c5e61573373babb5082331b98847a3cebf