Account resource requests of init containers
In before, we calculate the CPU/RAM of a capsule by summing
up the CPU/RAM of all its containers. This will over-calculate
the resource requirement of the capsule since init containers
won't run together with regular containers at any point in time.

This commit fixes this issue. The effective CPU/RAM of capsule
should be the maximum of:
* The largest CPU/RAM of init containers
* the sum of all regular containers

Change-Id: Ia3d4dec13d8cb8d32d2c0ae1ec714f00bcb9bc0a