prefix stdin with channel
k8s overlays a concept of "channels" on its websocket traffic. I'm not
sure how common of a thing this is, but it can be more or less supported
via the new websocket_opts function. If "channels" is defined there, channel
prefixes can be specified, and they'll be used to tag up the ws traffic.

Maybe should be chr()-encoded inside the options function instead of in the
websocket client.