Fix exec on k8s containers
Executing a command against a k8s container requires that the request
be in the form of a websocket. Additionally, getting the return code
is a bit trickier as you have to get a reference to the raw ws client
rather than the string response. Turns out the returncode accessor can
also throw for certain command inputs (like a bogus command binary that
can't be found.)