Fix compatibility with oslo.db 12.1.0
oslo.db 12.1.0 has changed the default value for the 'autocommit'
parameter of 'LegacyEngineFacade' from 'True' to 'False'. This is a
necessary step to ensure compatibility with SQLAlchemy 2.0. However, we
are currently relying on the autocommit behavior and need changes to
explicitly manage sessions. Until that happens, we need to override the
default.

Also declare queue at top level so that the tests actually run.

Change-Id: I04b336a42b5cbb458badde3bb04d35363d7834d8
(cherry picked from commit 9d92de5423a311630d250b21269067679f3e16f5)