Use alembic alter_column for 35cb52c5553f migration
With latest sqlalchemy there is no alter attribute for the Column object

Thus, zun-db-manage upgrade fails with AttributeError. Instead of trying
to use SQLAlchemy for altering Column we should use alembic instead
for the migration

This also fix py3.11 job by replacing libmysqlclient-dev with
libmariadb-dev-compat which is available for both ubuntu and debian.

Closes-Bug: #2043209
Change-Id: Iaba736d391ad5820de5f481d2c5bac1714dc0809